# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lattice API",
    "description": "Lattice Access Control System API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@yourorg.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/roles/{name}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get role by name",
        "description": "Retrieve a specific role by its name",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Role name"
          },
          {
            "in": "query",
            "name": "contextType",
            "schema": {
              "type": "string"
            },
            "description": "Context type"
          }
        ],
        "responses": {
          "200": {
            "description": "Role details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "description": "Role not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Delete role",
        "description": "Delete a role",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Role name"
          },
          {
            "in": "query",
            "name": "contextType",
            "schema": {
              "type": "string"
            },
            "description": "Context type"
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Role": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Role name"
          },
          "contextType": {
            "type": "string",
            "description": "Context type for this role"
          },
          "key": {
            "type": "string",
            "description": "Role key"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Validation issues"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "JWT Bearer token for authentication"
      }
    }
  }
}